refactor(team): phase3 migrate team domain models into crate
