+++ crates/agenthub-team-domain/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "agenthub-team-domain"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+thiserror = "1"
